Split stocks utils in many files
