+++ saleor/checkout/utils.py
@@ -44,11 +44,8 @@ from ..order.actions import order_created
 from ..order.emails import send_order_confirmation
 from ..order.models import Order, OrderLine
 from ..shipping.models import ShippingMethod
-from ..stock.stock_management import (
-    allocate_stock,
-    check_stock_quantity,
-    get_available_quantity,
-)
+from ..stock.utils.availability import check_stock_quantity, get_available_quantity
+from ..stock.utils.management import allocate_stock
 from . import AddressType, logger
 from .forms import (
     AddressChoiceForm,
